reorder instructions and add placeholders for new ones
