+++ assembly/instructions.ts
@@ -23,9 +23,8 @@ export const MISSING_INSTRUCTION = instruction("INVALID", Arguments.Zero, 1, fal
 export const SBRK = instruction("SBRK", Arguments.TwoReg, 1);
 
 export const INSTRUCTIONS: Instruction[] = [
-  // 0
-  instruction("TRAP", Arguments.Zero, 1, true),
-  instruction("FALLTHROUGH", Arguments.Zero, 1, true),
+  /* 000 */ instruction("TRAP", Arguments.Zero, 1, true),
+  /* 001 */ instruction("FALLTHROUGH", Arguments.Zero, 1, true),
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
@@ -35,8 +34,7 @@ export const INSTRUCTIONS: Instruction[] = [
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
 
-  // 10
-  instruction("ECALLI", Arguments.OneImm, 1),
+  /* 010 */ instruction("ECALLI", Arguments.OneImm, 1),
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
@@ -47,8 +45,7 @@ export const INSTRUCTIONS: Instruction[] = [
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
 
-  // 20
-  instruction("LOAD_IMM_64", Arguments.OneRegOneExtImm, 1),
+  /* 020 */ instruction("LOAD_IMM_64", Arguments.OneRegOneExtImm, 1),
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
@@ -59,11 +56,10 @@ export const INSTRUCTIONS: Instruction[] = [
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
 
-  // 30
-  instruction("STORE_IMM_U8", Arguments.TwoImm, 1),
-  instruction("STORE_IMM_U16", Arguments.TwoImm, 1),
-  instruction("STORE_IMM_U32", Arguments.TwoImm, 1),
-  instruction("STORE_IMM_U64", Arguments.TwoImm, 1),
+  /* 030 */ instruction("STORE_IMM_U8", Arguments.TwoImm, 1),
+  /* 031 */ instruction("STORE_IMM_U16", Arguments.TwoImm, 1),
+  /* 032 */ instruction("STORE_IMM_U32", Arguments.TwoImm, 1),
+  /* 033 */ instruction("STORE_IMM_U64", Arguments.TwoImm, 1),
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
@@ -71,8 +67,7 @@ export const INSTRUCTIONS: Instruction[] = [
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
 
-  // 40
-  instruction("JUMP", Arguments.OneOff, 1, true),
+  /* 040 */ instruction("JUMP", Arguments.OneOff, 1, true),
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
@@ -83,22 +78,20 @@ export const INSTRUCTIONS: Instruction[] = [
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
 
-  // 50
-  instruction("JUMP_IND", Arguments.OneRegOneImm, 1, true),
-  instruction("LOAD_IMM", Arguments.OneRegOneImm, 1),
-  instruction("LOAD_U8", Arguments.OneRegOneImm, 1),
-  instruction("LOAD_I8", Arguments.OneRegOneImm, 1),
-  instruction("LOAD_U16", Arguments.OneRegOneImm, 1),
-  instruction("LOAD_I16", Arguments.OneRegOneImm, 1),
-  instruction("LOAD_U32", Arguments.OneRegOneImm, 1),
-  instruction("LOAD_I32", Arguments.OneRegOneImm, 1),
-  instruction("LOAD_U64", Arguments.OneRegOneImm, 1),
-  instruction("STORE_U8", Arguments.OneRegOneImm, 1),
+  /* 050 */ instruction("JUMP_IND", Arguments.OneRegOneImm, 1, true),
+  /* 051 */ instruction("LOAD_IMM", Arguments.OneRegOneImm, 1),
+  /* 052 */ instruction("LOAD_U8", Arguments.OneRegOneImm, 1),
+  /* 053 */ instruction("LOAD_I8", Arguments.OneRegOneImm, 1),
+  /* 054 */ instruction("LOAD_U16", Arguments.OneRegOneImm, 1),
+  /* 055 */ instruction("LOAD_I16", Arguments.OneRegOneImm, 1),
+  /* 056 */ instruction("LOAD_U32", Arguments.OneRegOneImm, 1),
+  /* 057 */ instruction("LOAD_I32", Arguments.OneRegOneImm, 1),
+  /* 058 */ instruction("LOAD_U64", Arguments.OneRegOneImm, 1),
+  /* 059 */ instruction("STORE_U8", Arguments.OneRegOneImm, 1),
 
-  // 60
-  instruction("STORE_U16", Arguments.OneRegOneImm, 1),
-  instruction("STORE_U32", Arguments.OneRegOneImm, 1),
-  instruction("STORE_U64", Arguments.OneRegOneImm, 1),
+  /* 060 */ instruction("STORE_U16", Arguments.OneRegOneImm, 1),
+  /* 061 */ instruction("STORE_U32", Arguments.OneRegOneImm, 1),
+  /* 062 */ instruction("STORE_U64", Arguments.OneRegOneImm, 1),
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
@@ -107,11 +100,10 @@ export const INSTRUCTIONS: Instruction[] = [
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
 
-  // 70
-  instruction("STORE_IMM_IND_U8", Arguments.OneRegTwoImm, 1),
-  instruction("STORE_IMM_IND_U16", Arguments.OneRegTwoImm, 1),
-  instruction("STORE_IMM_IND_U32", Arguments.OneRegTwoImm, 1),
-  instruction("STORE_IMM_IND_U64", Arguments.OneRegTwoImm, 1),
+  /* 070 */ instruction("STORE_IMM_IND_U8", Arguments.OneRegTwoImm, 1),
+  /* 071 */ instruction("STORE_IMM_IND_U16", Arguments.OneRegTwoImm, 1),
+  /* 072 */ instruction("STORE_IMM_IND_U32", Arguments.OneRegTwoImm, 1),
+  /* 073 */ instruction("STORE_IMM_IND_U64", Arguments.OneRegTwoImm, 1),
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
@@ -119,20 +111,18 @@ export const INSTRUCTIONS: Instruction[] = [
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
 
-  // 80
-  instruction("LOAD_IMM_JUMP", Arguments.OneRegOneImmOneOff, 1, true),
-  instruction("BRANCH_EQ_IMM", Arguments.OneRegOneImmOneOff, 1, true),
-  instruction("BRANCH_NE_IMM", Arguments.OneRegOneImmOneOff, 1, true),
-  instruction("BRANCH_LT_U_IMM", Arguments.OneRegOneImmOneOff, 1, true),
-  instruction("BRANCH_LE_U_IMM", Arguments.OneRegOneImmOneOff, 1, true),
-  instruction("BRANCH_GE_U_IMM", Arguments.OneRegOneImmOneOff, 1, true),
-  instruction("BRANCH_GT_U_IMM", Arguments.OneRegOneImmOneOff, 1, true),
-  instruction("BRANCH_LT_S_IMM", Arguments.OneRegOneImmOneOff, 1, true),
-  instruction("BRANCH_LE_S_IMM", Arguments.OneRegOneImmOneOff, 1, true),
-  instruction("BRANCH_GE_S_IMM", Arguments.OneRegOneImmOneOff, 1, true),
+  /* 080 */ instruction("LOAD_IMM_JUMP", Arguments.OneRegOneImmOneOff, 1, true),
+  /* 081 */ instruction("BRANCH_EQ_IMM", Arguments.OneRegOneImmOneOff, 1, true),
+  /* 082 */ instruction("BRANCH_NE_IMM", Arguments.OneRegOneImmOneOff, 1, true),
+  /* 083 */ instruction("BRANCH_LT_U_IMM", Arguments.OneRegOneImmOneOff, 1, true),
+  /* 084 */ instruction("BRANCH_LE_U_IMM", Arguments.OneRegOneImmOneOff, 1, true),
+  /* 085 */ instruction("BRANCH_GE_U_IMM", Arguments.OneRegOneImmOneOff, 1, true),
+  /* 086 */ instruction("BRANCH_GT_U_IMM", Arguments.OneRegOneImmOneOff, 1, true),
+  /* 087 */ instruction("BRANCH_LT_S_IMM", Arguments.OneRegOneImmOneOff, 1, true),
+  /* 088 */ instruction("BRANCH_LE_S_IMM", Arguments.OneRegOneImmOneOff, 1, true),
+  /* 089 */ instruction("BRANCH_GE_S_IMM", Arguments.OneRegOneImmOneOff, 1, true),
 
-  // 90
-  instruction("BRANCH_GT_S_IMM", Arguments.OneRegOneImmOneOff, 1, true),
+  /* 090 */ instruction("BRANCH_GT_S_IMM", Arguments.OneRegOneImmOneOff, 1, true),
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
@@ -143,9 +133,21 @@ export const INSTRUCTIONS: Instruction[] = [
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
 
-  // 100
-  instruction("MOVE_REG", Arguments.TwoReg, 1),
-  SBRK,
+  /* 100 */ instruction("MOVE_REG", Arguments.TwoReg, 1),
+  /* 101 */ SBRK,
+  // new (start)
+  /* 102 */ instruction("COUNT_SET_BITS_64", Arguments.TwoReg, 1),
+  /* 103 */ instruction("COUNT_SET_BITS_32", Arguments.TwoReg, 1),
+  /* 104 */ instruction("LEADING_ZERO_BITS_64", Arguments.TwoReg, 1),
+  /* 105 */ instruction("LEADING_ZERO_BITS_32", Arguments.TwoReg, 1),
+  /* 106 */ instruction("TRAILING_ZERO_BITS_64", Arguments.TwoReg, 1),
+  /* 107 */ instruction("TRAILING_ZERO_BITS_32", Arguments.TwoReg, 1),
+  /* 108 */ instruction("SIGN_EXTEND_8", Arguments.TwoReg, 1),
+  /* 109 */ instruction("SIGN_EXTEND_16", Arguments.TwoReg, 1),
+
+  /* 110 */ instruction("ZERO_EXTEND_16", Arguments.TwoReg, 1),
+  /* 111 */ instruction("REVERSE_BYTES", Arguments.TwoReg, 1),
+  // new (end)
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
@@ -155,68 +157,54 @@ export const INSTRUCTIONS: Instruction[] = [
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
 
-  // 110
-  instruction("STORE_IND_U8", Arguments.TwoRegOneImm, 1),
-  instruction("STORE_IND_U16", Arguments.TwoRegOneImm, 1),
-  instruction("STORE_IND_U32", Arguments.TwoRegOneImm, 1),
-  instruction("STORE_IND_U64", Arguments.TwoRegOneImm, 1),
-  instruction("LOAD_IND_U8", Arguments.TwoRegOneImm, 1),
-  instruction("LOAD_IND_I8", Arguments.TwoRegOneImm, 1),
-  instruction("LOAD_IND_U16", Arguments.TwoRegOneImm, 1),
-  instruction("LOAD_IND_I16", Arguments.TwoRegOneImm, 1),
-  instruction("LOAD_IND_U32", Arguments.TwoRegOneImm, 1),
-  instruction("LOAD_IND_I32", Arguments.TwoRegOneImm, 1),
+  /* 120 */ instruction("STORE_IND_U8", Arguments.TwoRegOneImm, 1),
+  /* 121 */ instruction("STORE_IND_U16", Arguments.TwoRegOneImm, 1),
+  /* 122 */ instruction("STORE_IND_U32", Arguments.TwoRegOneImm, 1),
+  /* 123 */ instruction("STORE_IND_U64", Arguments.TwoRegOneImm, 1),
+  /* 124 */ instruction("LOAD_IND_U8", Arguments.TwoRegOneImm, 1),
+  /* 125 */ instruction("LOAD_IND_I8", Arguments.TwoRegOneImm, 1),
+  /* 126 */ instruction("LOAD_IND_U16", Arguments.TwoRegOneImm, 1),
+  /* 127 */ instruction("LOAD_IND_I16", Arguments.TwoRegOneImm, 1),
+  /* 128 */ instruction("LOAD_IND_U32", Arguments.TwoRegOneImm, 1),
+  /* 129 */ instruction("LOAD_IND_I32", Arguments.TwoRegOneImm, 1),
 
-  // 120
-  instruction("LOAD_IND_U64", Arguments.TwoRegOneImm, 1),
-  instruction("ADD_IMM_32", Arguments.TwoRegOneImm, 1),
-  instruction("AND_IMM", Arguments.TwoRegOneImm, 1),
-  instruction("XOR_IMM", Arguments.TwoRegOneImm, 1),
-  instruction("OR_IMM", Arguments.TwoRegOneImm, 1),
-  instruction("MUL_IMM_32", Arguments.TwoRegOneImm, 1),
-  instruction("SET_LT_U_IMM", Arguments.TwoRegOneImm, 1),
-  instruction("SET_LT_S_IMM", Arguments.TwoRegOneImm, 1),
-  instruction("SHLO_L_IMM_32", Arguments.TwoRegOneImm, 1),
-  instruction("SHLO_R_IMM_32", Arguments.TwoRegOneImm, 1),
+  /* 130 */ instruction("LOAD_IND_U64", Arguments.TwoRegOneImm, 1),
+  /* 131 */ instruction("ADD_IMM_32", Arguments.TwoRegOneImm, 1),
+  /* 132 */ instruction("AND_IMM", Arguments.TwoRegOneImm, 1),
+  /* 133 */ instruction("XOR_IMM", Arguments.TwoRegOneImm, 1),
+  /* 134 */ instruction("OR_IMM", Arguments.TwoRegOneImm, 1),
+  /* 135 */ instruction("MUL_IMM_32", Arguments.TwoRegOneImm, 1),
+  /* 136 */ instruction("SET_LT_U_IMM", Arguments.TwoRegOneImm, 1),
+  /* 137 */ instruction("SET_LT_S_IMM", Arguments.TwoRegOneImm, 1),
+  /* 138 */ instruction("SHLO_L_IMM_32", Arguments.TwoRegOneImm, 1),
+  /* 139 */ instruction("SHLO_R_IMM_32", Arguments.TwoRegOneImm, 1),
 
-  // 130
-  instruction("SHAR_R_IMM_32", Arguments.TwoRegOneImm, 1),
-  instruction("NEG_ADD_IMM_32", Arguments.TwoRegOneImm, 1),
-  instruction("SET_GT_U_IMM", Arguments.TwoRegOneImm, 1),
-  instruction("SET_GT_S_IMM", Arguments.TwoRegOneImm, 1),
-  instruction("SHLO_L_IMM_ALT_32", Arguments.TwoRegOneImm, 1),
-  instruction("SHLO_R_IMM_ALT_32", Arguments.TwoRegOneImm, 1),
-  instruction("SHAR_R_IMM_ALT_32", Arguments.TwoRegOneImm, 1),
-  instruction("CMOV_IZ_IMM", Arguments.TwoRegOneImm, 1),
-  instruction("CMOV_NZ_IMM", Arguments.TwoRegOneImm, 1),
-  instruction("ADD_IMM_64", Arguments.TwoRegOneImm, 1),
+  /* 140 */ instruction("SHAR_R_IMM_32", Arguments.TwoRegOneImm, 1),
+  /* 141 */ instruction("NEG_ADD_IMM_32", Arguments.TwoRegOneImm, 1),
+  /* 142 */ instruction("SET_GT_U_IMM", Arguments.TwoRegOneImm, 1),
+  /* 143 */ instruction("SET_GT_S_IMM", Arguments.TwoRegOneImm, 1),
+  /* 144 */ instruction("SHLO_L_IMM_ALT_32", Arguments.TwoRegOneImm, 1),
+  /* 145 */ instruction("SHLO_R_IMM_ALT_32", Arguments.TwoRegOneImm, 1),
+  /* 146 */ instruction("SHAR_R_IMM_ALT_32", Arguments.TwoRegOneImm, 1),
+  /* 147 */ instruction("CMOV_IZ_IMM", Arguments.TwoRegOneImm, 1),
+  /* 148 */ instruction("CMOV_NZ_IMM", Arguments.TwoRegOneImm, 1),
+  /* 149 */ instruction("ADD_IMM_64", Arguments.TwoRegOneImm, 1),
 
-  // 140
-  instruction("MUL_IMM_64", Arguments.TwoRegOneImm, 1),
-  instruction("SHLO_L_IMM_64", Arguments.TwoRegOneImm, 1),
-  instruction("SHLO_R_IMM_64", Arguments.TwoRegOneImm, 1),
-  instruction("SHAR_R_IMM_64", Arguments.TwoRegOneImm, 1),
-  instruction("NEG_ADD_IMM_64", Arguments.TwoRegOneImm, 1),
-  instruction("SHLO_L_IMM_ALT_64", Arguments.TwoRegOneImm, 1),
-  instruction("SHLO_R_IMM_ALT_64", Arguments.TwoRegOneImm, 1),
-  instruction("SHAR_R_IMM_ALT_64", Arguments.TwoRegOneImm, 1),
-  MISSING_INSTRUCTION,
-  MISSING_INSTRUCTION,
+  /* 150 */ instruction("MUL_IMM_64", Arguments.TwoRegOneImm, 1),
+  /* 151 */ instruction("SHLO_L_IMM_64", Arguments.TwoRegOneImm, 1),
+  /* 152 */ instruction("SHLO_R_IMM_64", Arguments.TwoRegOneImm, 1),
+  /* 153 */ instruction("SHAR_R_IMM_64", Arguments.TwoRegOneImm, 1),
+  /* 154 */ instruction("NEG_ADD_IMM_64", Arguments.TwoRegOneImm, 1),
+  /* 155 */ instruction("SHLO_L_IMM_ALT_64", Arguments.TwoRegOneImm, 1),
+  /* 156 */ instruction("SHLO_R_IMM_ALT_64", Arguments.TwoRegOneImm, 1),
+  /* 157 */ instruction("SHAR_R_IMM_ALT_64", Arguments.TwoRegOneImm, 1),
+  // new (start)
+  /* 158 */ instruction("ROT_R_64_IMM", Arguments.TwoRegOneImm, 1),
+  /* 159 */ instruction("ROT_R_64_IMM_ALT", Arguments.TwoRegOneImm, 1),
 
-  // 150
-  instruction("BRANCH_EQ", Arguments.TwoRegOneOff, 1, true),
-  instruction("BRANCH_NE", Arguments.TwoRegOneOff, 1, true),
-  instruction("BRANCH_LT_U", Arguments.TwoRegOneOff, 1, true),
-  instruction("BRANCH_LT_S", Arguments.TwoRegOneOff, 1, true),
-  instruction("BRANCH_GE_U", Arguments.TwoRegOneOff, 1, true),
-  instruction("BRANCH_GE_S", Arguments.TwoRegOneOff, 1, true),
-  MISSING_INSTRUCTION,
-  MISSING_INSTRUCTION,
-  MISSING_INSTRUCTION,
-  MISSING_INSTRUCTION,
-
-  // 160
-  instruction("LOAD_IMM_JUMP_IND", Arguments.TwoRegTwoImm, 1, true),
+  /* 160 */ instruction("ROT_R_32_IMM", Arguments.TwoRegOneImm, 1),
+  /* 161 */ instruction("ROT_R_32_IMM_ALT", Arguments.TwoRegOneImm, 1),
+  // new (end)
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
@@ -227,39 +215,73 @@ export const INSTRUCTIONS: Instruction[] = [
   MISSING_INSTRUCTION,
   MISSING_INSTRUCTION,
 
-  // 170
-  instruction("ADD_32", Arguments.ThreeReg, 1),
-  instruction("SUB_32", Arguments.ThreeReg, 1),
-  instruction("MUL_32", Arguments.ThreeReg, 1),
-  instruction("DIV_U_32", Arguments.ThreeReg, 1),
-  instruction("DIV_S_32", Arguments.ThreeReg, 1),
-  instruction("REM_U_32", Arguments.ThreeReg, 1),
-  instruction("REM_S_32", Arguments.ThreeReg, 1),
-  instruction("SHLO_L_32", Arguments.ThreeReg, 1),
-  instruction("SHLO_R_32", Arguments.ThreeReg, 1),
-  instruction("SHAR_R_32", Arguments.ThreeReg, 1),
+  /* 170 */ instruction("BRANCH_EQ", Arguments.TwoRegOneOff, 1, true),
+  /* 171 */ instruction("BRANCH_NE", Arguments.TwoRegOneOff, 1, true),
+  /* 172 */ instruction("BRANCH_LT_U", Arguments.TwoRegOneOff, 1, true),
+  /* 173 */ instruction("BRANCH_LT_S", Arguments.TwoRegOneOff, 1, true),
+  /* 174 */ instruction("BRANCH_GE_U", Arguments.TwoRegOneOff, 1, true),
+  /* 175 */ instruction("BRANCH_GE_S", Arguments.TwoRegOneOff, 1, true),
+  MISSING_INSTRUCTION,
+  MISSING_INSTRUCTION,
+  MISSING_INSTRUCTION,
+  MISSING_INSTRUCTION,
 
-  // 180
-  instruction("ADD_64", Arguments.ThreeReg, 1),
-  instruction("SUB_64", Arguments.ThreeReg, 1),
-  instruction("MUL_64", Arguments.ThreeReg, 1),
-  instruction("DIV_U_64", Arguments.ThreeReg, 1),
-  instruction("DIV_S_64", Arguments.ThreeReg, 1),
-  instruction("REM_U_64", Arguments.ThreeReg, 1),
-  instruction("REM_S_64", Arguments.ThreeReg, 1),
-  instruction("SHLO_L_64", Arguments.ThreeReg, 1),
-  instruction("SHLO_R_64", Arguments.ThreeReg, 1),
-  instruction("SHAR_R_64", Arguments.ThreeReg, 1),
+  /* 180 */ instruction("LOAD_IMM_JUMP_IND", Arguments.TwoRegTwoImm, 1, true),
+  MISSING_INSTRUCTION,
+  MISSING_INSTRUCTION,
+  MISSING_INSTRUCTION,
+  MISSING_INSTRUCTION,
+  MISSING_INSTRUCTION,
+  MISSING_INSTRUCTION,
+  MISSING_INSTRUCTION,
+  MISSING_INSTRUCTION,
+  MISSING_INSTRUCTION,
 
-  // 190
-  instruction("AND", Arguments.ThreeReg, 1),
-  instruction("XOR", Arguments.ThreeReg, 1),
-  instruction("OR", Arguments.ThreeReg, 1),
-  instruction("MUL_UPPER_S_S", Arguments.ThreeReg, 1),
-  instruction("MUL_UPPER_U_U", Arguments.ThreeReg, 1),
-  instruction("MUL_UPPER_S_U", Arguments.ThreeReg, 1),
-  instruction("SET_LT_U", Arguments.ThreeReg, 1),
-  instruction("SET_LT_S", Arguments.ThreeReg, 1),
-  instruction("CMOV_IZ", Arguments.ThreeReg, 1),
-  instruction("CMOV_NZ", Arguments.ThreeReg, 1),
+  /* 190 */ instruction("ADD_32", Arguments.ThreeReg, 1),
+  /* 191 */ instruction("SUB_32", Arguments.ThreeReg, 1),
+  /* 192 */ instruction("MUL_32", Arguments.ThreeReg, 1),
+  /* 193 */ instruction("DIV_U_32", Arguments.ThreeReg, 1),
+  /* 194 */ instruction("DIV_S_32", Arguments.ThreeReg, 1),
+  /* 195 */ instruction("REM_U_32", Arguments.ThreeReg, 1),
+  /* 196 */ instruction("REM_S_32", Arguments.ThreeReg, 1),
+  /* 197 */ instruction("SHLO_L_32", Arguments.ThreeReg, 1),
+  /* 198 */ instruction("SHLO_R_32", Arguments.ThreeReg, 1),
+  /* 199 */ instruction("SHAR_R_32", Arguments.ThreeReg, 1),
+
+  /* 200 */ instruction("ADD_64", Arguments.ThreeReg, 1),
+  /* 201 */ instruction("SUB_64", Arguments.ThreeReg, 1),
+  /* 202 */ instruction("MUL_64", Arguments.ThreeReg, 1),
+  /* 203 */ instruction("DIV_U_64", Arguments.ThreeReg, 1),
+  /* 204 */ instruction("DIV_S_64", Arguments.ThreeReg, 1),
+  /* 205 */ instruction("REM_U_64", Arguments.ThreeReg, 1),
+  /* 206 */ instruction("REM_S_64", Arguments.ThreeReg, 1),
+  /* 207 */ instruction("SHLO_L_64", Arguments.ThreeReg, 1),
+  /* 208 */ instruction("SHLO_R_64", Arguments.ThreeReg, 1),
+  /* 209 */ instruction("SHAR_R_64", Arguments.ThreeReg, 1),
+
+  /* 210 */ instruction("AND", Arguments.ThreeReg, 1),
+  /* 211 */ instruction("XOR", Arguments.ThreeReg, 1),
+  /* 212 */ instruction("OR", Arguments.ThreeReg, 1),
+  /* 213 */ instruction("MUL_UPPER_S_S", Arguments.ThreeReg, 1),
+  /* 214 */ instruction("MUL_UPPER_U_U", Arguments.ThreeReg, 1),
+  /* 215 */ instruction("MUL_UPPER_S_U", Arguments.ThreeReg, 1),
+  /* 216 */ instruction("SET_LT_U", Arguments.ThreeReg, 1),
+  /* 217 */ instruction("SET_LT_S", Arguments.ThreeReg, 1),
+  /* 218 */ instruction("CMOV_IZ", Arguments.ThreeReg, 1),
+  /* 219 */ instruction("CMOV_NZ", Arguments.ThreeReg, 1),
+
+  // new (start)
+  /* 220 */ instruction("ROT_L_64", Arguments.ThreeReg, 1),
+  /* 221 */ instruction("ROT_L_32", Arguments.ThreeReg, 1),
+  /* 222 */ instruction("ROT_R_64", Arguments.ThreeReg, 1),
+  /* 223 */ instruction("ROT_R_32", Arguments.ThreeReg, 1),
+  /* 224 */ instruction("AND_INV", Arguments.ThreeReg, 1),
+  /* 225 */ instruction("OR_INV", Arguments.ThreeReg, 1),
+  /* 226 */ instruction("XNOR", Arguments.ThreeReg, 1),
+  /* 227 */ instruction("MAX", Arguments.ThreeReg, 1),
+  /* 228 */ instruction("MAX_U", Arguments.ThreeReg, 1),
+  /* 229 */ instruction("MIN", Arguments.ThreeReg, 1),
+
+  /* 230 */ instruction("MIN_U", Arguments.ThreeReg, 1),
+  // new (end)
 ];
