+++ assembly/instructions/bit.ts
@@ -1,0 +1,31 @@
+import {InstructionRun, ok} from "./outcome";
+
+// COUNT_SET_BITS_64
+export const count_set_bits_64: InstructionRun = () => ok();
+
+// COUNT_SET_BITS_32
+export const count_set_bits_32: InstructionRun = () => ok();
+
+// LEADING_ZERO_BITS_64
+export const leading_zero_bits_64: InstructionRun = () => ok();
+
+// LEADING_ZERO_BITS_32
+export const leading_zero_bits_32: InstructionRun = () => ok();
+
+// TRAILING_ZERO_BITS_64
+export const trailing_zero_bits_64: InstructionRun = () => ok();
+
+// TRAILING_ZERO_BITS_32
+export const trailing_zero_bits_32: InstructionRun = () => ok();
+
+// SIGN_EXTEND_8
+export const sign_extend_8: InstructionRun = () => ok();
+
+// SIGN_EXTEND_16
+export const sign_extend_16: InstructionRun = () => ok();
+
+// ZERO_EXTEND_16
+export const zero_extend_16: InstructionRun = () => ok();
+
+// REVERSE_BYTES
+export const reverse_bytes: InstructionRun = () => ok();
