+++ assembly/instructions/rot.ts
@@ -1,0 +1,26 @@
+import {InstructionRun, ok} from "./outcome";
+
+// ROT_R_64_IMM
+export const rot_r_64_imm: InstructionRun = () => ok();
+
+// ROT_R_64_IMM_ALT
+export const rot_r_64_imm_alt: InstructionRun = () => ok();
+
+// ROT_R_32_IMM
+export const rot_r_32_imm: InstructionRun = () => ok();
+
+// ROT_R_32_IMM_ALT
+export const rot_r_32_imm_alt: InstructionRun = () => ok();
+
+// ROT_L_64
+export const rot_l_64: InstructionRun = () => ok();
+
+// ROT_L_32
+export const rot_l_32: InstructionRun = () => ok();
+
+// ROT_R_64
+export const rot_r_64: InstructionRun = () => ok();
+
+// ROT_R_32
+export const rot_r_32: InstructionRun = () => ok();
+
